Calculate point estimates and errors

Calculates point estimates and errors using the package `survey` based on sampling weights from a survey design

calcPtEst(df, var, use_weights = FALSE, by_var = NA, design = NA,
  psu_var = NA, strata_var = NA, weight_var = NA, na.rm = TRUE,
  ci_factor = 2)

Arguments

df
main data frame containing the raw data

var
string containing the variable name you want to average

use_weights
TRUE/FALSE for whether to apply a sampling frame or calculate a simple average.

by_var
(optional) string containing the variable name over which you want to calculate the estimates

design
(optional) svydesign object containing the sample frame

psu_var
(optional) if design isn't specified, string containing the primary sampling unit variable from the survey design (argument id in library('survey'))

strata_var
(optional) if design isn't specified, string containing the strata variable from the survey design (argument strata in library('survey')

weight_var
(optional)if design isn't specified, string containing the weights variable from the survey design (argument weights in library('survey')

na.rm
remove NAs from the mean or not

ci_factor
value to calculate confidence interval; in standard deviations. 1.96 standard deviations --> ~ 95 percent of the area under a normal gaussian distribution

Examples

# Generate simple random data to average
df = data.frame(region = rep(letters, 100), value = sample(100, 2600, replace = TRUE))

calcPtEst(df, 'value', by_var = 'region', use_weights = FALSE)

#> # A tibble: 26 x 7
#>    region     N unweighted_avg unweighted_sd       ci       ub       lb
#>    <fctr> <int>          <dbl>         <dbl>    <dbl>    <dbl>    <dbl>
#>  1      o   100          55.85      28.72031 5.744061 61.59406 50.10594
#>  2      y   100          54.49      27.07099 5.414197 59.90420 49.07580
#>  3      b   100          53.92      29.90344 5.980688 59.90069 47.93931
#>  4      e   100          53.79      27.96581 5.593163 59.38316 48.19684
#>  5      t   100          53.55      30.44564 6.089128 59.63913 47.46087
#>  6      r   100          53.31      29.17024 5.834048 59.14405 47.47595
#>  7      x   100          52.99      27.87291 5.574581 58.56458 47.41542
#>  8      d   100          52.34      29.30723 5.861445 58.20145 46.47855
#>  9      l   100          51.76      30.14581 6.029163 57.78916 45.73084
#> 10      v   100          51.75      29.66015 5.932031 57.68203 45.81797
#> # ... with 16 more rows

# generate data from a survey
data(api, package = 'survey')
dclus1 <- survey::svydesign(id=~dnum, weights=~pw, data=apiclus1, fpc=~fpc)
# no weights
calcPtEst(apiclus1, var = 'api99', by_var = 'stype', use_weights = FALSE)

#> # A tibble: 3 x 7
#>    stype     N unweighted_avg unweighted_sd       ci       ub       lb
#>   <fctr> <int>          <dbl>         <dbl>    <dbl>    <dbl>    <dbl>
#> 1      M    25       608.6000      110.7429 44.29718 652.8972 564.3028
#> 2      E   144       607.7917      114.4657 19.07762 626.8693 588.7140
#> 3      H    14       595.7143      106.6594 57.01183 652.7261 538.7025

# specifying the sampling frame as a design object
calcPtEst(apiclus1, var = 'api99', by_var = 'stype', design = dclus1, use_weights = TRUE)

#> Warning: the condition has length > 1 and only the first element will be used

#> Joining, by = "stype"

#>   stype    api99       se      avg       ci       ub       lb   N
#> 1     M 608.6000 32.56064 608.6000 65.12128 673.7213 543.4787  25
#> 2     E 607.7917 22.81660 607.7917 45.63320 653.4249 562.1585 144
#> 3     H 595.7143 41.76400 595.7143 83.52800 679.2423 512.1863  14
#>   unweighted_avg unweighted_sd
#> 1       608.6000      110.7429
#> 2       607.7917      114.4657
#> 3       595.7143      106.6594