Create a lookup table for DHS country and indicator codes, and find the country/indicator code based on the list

Create a lookup table for DHS country and indicator codes, and find the country/indicator code based on the list

getDHScountry(country_names, return_table = FALSE)

getDHSindicator(indicators, refresh_indicators = FALSE,
  return_table = FALSE)

Arguments

country_names
string or list of strings of country names

return_table
whether to return the codes in a table containing their name and code, or just a collapsed list of the codes

indicators
string or list of strings of indicator names (or parts of their names)

refresh_indicators
T/F of whether to re-pull the indicator list from the DHS website. Takes a bit of time to execute

Functions

getDHScountry: Function to pull the DHS codes for country names. Can take either a single country or multiple ones; also can return entire table or just the country code Matching is *exact* matching; input name must exactly match a country name within the DHS dataset.

getDHSindicator: Function to lookup the DHS indicator codes. Partially matches to input indicator; user can select which indicator(s) they want

Examples

getDHScountry(c('Niger', 'Nigeria', 'Rwanda', 'Zambia'), return_table = FALSE)

#> [1] "NI,NG,RW,ZM"

getDHScountry(c('Niger', 'Nigeria', 'Rwanda', 'Zambia'), return_table = TRUE)

#> # A tibble: 4 x 2
#>    code country
#>   <chr>   <chr>
#> 1    NI   Niger
#> 2    NG Nigeria
#> 3    RW  Rwanda
#> 4    ZM  Zambia

getDHScountry('Rwanda')

#> [1] "RW"

getDHScountry('Nige') # Returns an error.

#> Warning: No countries found. Country names include: Afghanistan, Albania, Angola, Armenia, Azerbaijan, Bangladesh, Benin, Bolivia, Botswana, Brazil, Burkina Faso, Burundi, Cambodia, Cameroon, Cape Verde, Central African Republic, Chad, Colombia, Comoros, Congo, Congo Democratic Republic, Cote d'Ivoire, Dominican Republic, Ecuador, Egypt, El Salvador, Equatorial Guinea, Eritrea, Ethiopia, Gabon, Gambia, Ghana, Guatemala, Guinea, Guyana, Haiti, Honduras, India, Indonesia, Jordan, Kazakhstan, Kenya, Kyrgyz Republic, Lao People's Democratic Republic, Lesotho, , Liberia, Madagascar, Malawi, Maldives, Mali, Mauritania, Mexico, Moldova, Morocco, Mozambique, Myanmar, Namibia, Nepal, Nicaragua, Niger, Nigeria, Nigeria (Ondo State), Pakistan, Paraguay, Peru, Philippines, Rwanda, Samoa, Sao Tome and Principe, Senegal, Sierra Leone, South Africa, Sri Lanka, Sudan, Swaziland, Tajikistan, Tanzania, Thailand, Timor-Leste, Togo, Trinidad and Tobago, Tunisia, Turkey, Turkmenistan, Uganda, Ukraine, Uzbekistan, Vietnam, Yemen, Zambia, Zimbabwe

#> NULL

stunted_indic = getDHSindicator('stunted', return_table = TRUE)

#> Error in select.list(choices = c("all", as.character(filtered_indic$Label)),     multiple = TRUE, title = "Multiple indicators were found. Enter which one(s) you want."): select.list() cannot be used non-interactively

malnourished = getDHSindicator(c('stunted', 'wasted', 'underweight'))

#> Error in select.list(choices = c("all", as.character(filtered_indic$Label)),     multiple = TRUE, title = "Multiple indicators were found. Enter which one(s) you want."): select.list() cannot be used non-interactively

getDHSindicator('studnted') # returns nothing

#> Warning: No indicators found.

#> NULL